Skip clearing guarded DIMMs
The commit modifies script to skip clearing operation status for
the DIMMs only if the DIMM is guarded.

The change has been done as operational status was being cleared
for all the DIMMs at start up which should not be a case in case
the DIMM has been marked as guarded.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>