Reset operational status at start up
The commit removes special handling w.r.t. re-setting of
operational status for some of the FRUs at start up.
This is done considering that information related to guard
against core and dimms are not available at this point.

To cover the situation where LEDs needs to be light up for
guarded FRUs a different script will be trigerred once the
service holding guard informationis up.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>