Notify PIM to update functional property
During power-on, clear fault-leds scripts clears all the fault
leds by setting the functional property of the corresponding led FRU
to true. While setting the property, it calls the busctl set-property
method, which updates the data only on DBus and data is not persisted.

When BMC is rebooted while host is running, on system come back, the
stale data of the FRU which has functional as false will get updated on
the DBus, which marks that FRU as critical on ASMI.

Signed-off-by: RekhaAparna01 <vrekhaaparna@ibm.com>