clear cxp_port at chassis on (#46)
Initially, the script was by-passing PCIe cards for clearing fault
LEDs. As a consequence, it could also skip clearing cxp_port fault
leds, resulting it permanent persistence of cxp_port fault leds once
any error with cxp callout is logged in the system.
Removal of cxp_port from the by-passed fru list will avoid any such
future instance, where as in case there is already such situation in
the system, then to clear that without bringing the system down, a
change is added when chassis is on.

Signed-off-by: Sunny Srivastava <sunnsr25@in.ibm.com>